requirements.txt
requests 
pandas 
sqlalchemy 
psycopg2 
pyyaml